requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.10.0
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
centrifuge-python==0.4.2
certifi==2025.11.12
charset-normalizer==3.4.4
decorator==5.2.1
Django==5.2.7
django-redis==6.0.0
exceptiongroup==1.3.1
executing==2.2.1
Faker==37.12.0
frozenlist==1.8.0
gunicorn==23.0.0
idna==3.11
ipython==8.37.0
jedi==0.19.2
matplotlib-inline==0.2.1
multidict==6.7.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psycopg2==2.9.11
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
redis==7.1.0
requests==2.32.5
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
